Minor: change zip eocd size to int

+++ lib/src/main/kotlin/org/jetbrains/zip/signer/verifier/ZipVerifier.kt
@@ -47,7 +47,7 @@ object ZipVerifier {
         )
         val endOfCentralDirectory = dataSource.slice(
             zipSections.endOfCentralDirectoryOffset,
-            zipSections.endOfCentralDirectorySizeBytes
+            zipSections.endOfCentralDirectorySizeBytes.toLong()
         )
         return verify(beforeApkSigningBlock, signingBlock, centralDir, endOfCentralDirectory)
     }
